+++ polaris-server.yml
@@ -167,5 +167,7 @@ logging:
 # Limits the size of request bodies sent to Polaris. -1 means no limit.
 maxRequestBodyBytes: -1
 
-rateLimiting:
-  factoryType: no-op
+rateLimiter:
+  constructionTimeoutMillis: 2000
+  factory:
+    type: no-op
