+++ polaris-service/src/main/java/org/apache/polaris/service/ratelimiter/OpenTelemetryClock.java
@@ -16,48 +16,23 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.service.ratelimiting;
+package org.apache.polaris.service.ratelimiter;
 
 import io.opentelemetry.sdk.common.Clock;
 
 /** Implementation of the opentelemetry Clock interface using the Polaris Clock interface */
 public class OpenTelemetryClock implements Clock {
-  org.apache.polaris.service.ratelimiting.Clock clock;
+  org.apache.polaris.service.ratelimiter.Clock clock;
 
-  public OpenTelemetryClock(org.apache.polaris.service.ratelimiting.Clock clock) {
+  public OpenTelemetryClock(org.apache.polaris.service.ratelimiter.Clock clock) {
     this.clock = clock;
   }
 
-  /**
-   * Returns the current epoch timestamp in nanos from this clock. This timestamp should only be
-   * used to compute a current time. To compute a duration, timestamps should always be obtained
-   * using {@link #nanoTime()}. For example, this usage is correct.
-   *
-   * <pre>{@code
-   * long startNanos = clock.nanoTime();
-   * // Spend time...
-   * long durationNanos = clock.nanoTime() - startNanos;
-   * }</pre>
-   *
-   * <p>This usage is NOT correct.
-   *
-   * <pre>{@code
-   * long startNanos = clock.now();
-   * // Spend time...
-   * long durationNanos = clock.now() - startNanos;
-   * }</pre>
-   *
-   * <p>Calling this is equivalent to calling {@link #now(boolean)} with {@code highPrecision=true}.
-   */
   @Override
   public long now() {
     return clock.nanoTime();
   }
 
-  /**
-   * Returns a time measurement with nanosecond precision that can only be used to calculate elapsed
-   * time.
-   */
   @Override
   public long nanoTime() {
     return clock.nanoTime();
