+++ polaris-service/src/main/java/org/apache/polaris/service/ratelimiter/RateLimiter.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.service.ratelimiting;
+package org.apache.polaris.service.ratelimiter;
 
 /** Interface for rate limiting requests */
 public interface RateLimiter {
