+++ polaris-service/src/main/resources/META-INF/services/io.dropwizard.jackson.Discoverable
@@ -24,4 +24,4 @@ org.apache.polaris.service.context.RealmContextResolver
 org.apache.polaris.service.context.CallContextResolver
 org.apache.polaris.service.auth.TokenBrokerFactory
 org.apache.polaris.service.catalog.FileIOFactory
-org.apache.polaris.service.ratelimiting.RateLimiterFactory
+org.apache.polaris.service.ratelimiter.RateLimiterFactory
