+++ polaris-service/src/test/java/org/apache/polaris/service/ratelimiter/AsyncFallbackTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.service.ratelimiting;
+package org.apache.polaris.service.ratelimiter;
 
 import io.dropwizard.testing.ConfigOverride;
 import io.dropwizard.testing.ResourceHelpers;
@@ -27,21 +27,23 @@ import java.util.function.Consumer;
 import org.apache.polaris.service.PolarisApplication;
 import org.apache.polaris.service.PolarisApplicationIntegrationTest;
 import org.apache.polaris.service.config.PolarisApplicationConfig;
-import org.apache.polaris.service.test.MockClock;
-import org.apache.polaris.service.test.MockRateLimiterFactory;
 import org.apache.polaris.service.test.PolarisConnectionExtension;
 import org.apache.polaris.service.test.SnowmanCredentialsExtension;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-/** Main integration tests for rate limiting */
+/**
+ * Integration test that verifies the timeout behavior for fetching async rate limiters. This is in
+ * its own test class because the Dropwizard app is per test class and there isn't a great way to
+ * allow tests to clear the rate limiter cache.
+ */
 @ExtendWith({
   DropwizardExtensionsSupport.class,
   PolarisConnectionExtension.class,
   SnowmanCredentialsExtension.class
 })
-public class RateLimitingTest {
+public class AsyncFallbackTest {
   private static final DropwizardAppExtension<PolarisApplicationConfig> EXT =
       new DropwizardAppExtension<>(
           PolarisApplication.class,
@@ -49,7 +51,11 @@ public class RateLimitingTest {
           ConfigOverride.config(
               "server.applicationConnectors[0].port",
               "0"), // Bind to random port to support parallelism
-          ConfigOverride.config("server.adminConnectors[0].port", "0"));
+          ConfigOverride.config("server.adminConnectors[0].port", "0"),
+          ConfigOverride.config("rateLimiter.factory.type", "mock"),
+          ConfigOverride.config("rateLimiter.factory.requestsPerSecond", "0"),
+          ConfigOverride.config("rateLimiter.factory.windowSeconds", "0"),
+          ConfigOverride.config("rateLimiter.factory.delaySeconds", "999"));
 
   private static String userToken;
   private static String realm;
@@ -57,24 +63,13 @@ public class RateLimitingTest {
   @BeforeAll
   public static void setup(PolarisConnectionExtension.PolarisToken userToken) {
     realm = PolarisConnectionExtension.getTestRealm(PolarisApplicationIntegrationTest.class);
-    RateLimitingTest.userToken = userToken.token();
+    AsyncFallbackTest.userToken = userToken.token();
   }
 
   @Test
-  public void testRateLimiter() {
+  public void testRequestNotLimitedBecauseConstructionShouldTimeOut() {
     Consumer<Response.Status> requestAsserter =
         TestUtil.constructRequestAsserter(EXT, userToken, realm);
-    MockRateLimiterFactory factory =
-        (MockRateLimiterFactory) (EXT.getConfiguration().getRateLimiterFactory());
-    long windowMillis = (long) (factory.windowSeconds * 1000);
-
-    MockClock clock = MockRateLimiterFactory.clock;
-    clock.setMillis(2 * windowMillis); // Clear any counters from before this test
-    for (int i = 0; i < factory.requestsPerSecond * factory.windowSeconds; i++) {
-      requestAsserter.accept(Response.Status.OK);
-    }
-    requestAsserter.accept(Response.Status.TOO_MANY_REQUESTS);
-
-    clock.setMillis(4 * windowMillis); // Clear any counters from this test
+    requestAsserter.accept(Response.Status.OK);
   }
 }
