+++ polaris-service/src/test/resources/META-INF/services/org.apache.polaris.service.ratelimiter.RateLimiterFactory
@@ -17,4 +17,4 @@
 # under the License.
 #
 
-org.apache.polaris.service.ratelimiting.DefaultRateLimiterFactory
+org.apache.polaris.service.ratelimiter.MockRateLimiterFactory
